ref: move msw user handlers from auth to new user file
