+++ frontend/src/mocks/msw/handlers/index.ts
@@ -1,3 +1,4 @@
 import { authHandlers } from './auth'
+import { userHandlers } from './user'
 
-export const handlers = [...authHandlers]
+export const handlers = [...authHandlers, ...userHandlers]
